requirements.txt
antiorm==1.2.1
bottle==0.12.25
chardet==5.2.0
db==0.1.1
db-sqlite3==0.0.1
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
proxy_tools==0.1.0
pylibftdi==0.22.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pywebview==5.1
reportlab==4.2.2
scipy==1.14.0
typing_extensions==4.12.2
websockets==12.0
